Refactoring Tests code and other CR comments
