+++ src/WebJobs.Script/Description/Proxies/ProxyFunctionDescriptorProvider.cs
