+++ src/WebJobs.Script/Description/Proxies/ProxyFunctionInvoker.cs
