+++ src/WebJobs.Script/ScriptConstants.cs
@@ -9,7 +9,6 @@ namespace Microsoft.Azure.WebJobs.Script
     {
         public const string AzureFunctionsWebHookContextKey = "MS_AzureFunctionsWebHookContext";
         public const string AzureFunctionsHttpResponseKey = "MS_AzureFunctionsHttpResponse";
-        public const string AzureFunctionsProxyHttpRequestKey = "MS_AzureFunctionsHttpRequest";
         public const string AzureFunctionsHttpRequestAuthorizationLevelKey = "MS_AzureFunctionsAuthorizationLevel";
         public const string AzureFunctionsHttpRequestKeyNameKey = "MS_AzureFunctionsKeyId";
         public const string AzureFunctionsHttpRequestAuthorizationDisabledKey = "MS_AzureFunctionsAuthorizationDisabled";
