+++ test/WebJobs.Script.Tests.Integration/TestScripts/Proxies/host.json
