+++ test/WebJobs.Script.Tests.Integration/TestScripts/Proxies/httptrigger/function.json
