+++ test/WebJobs.Script.Tests.Integration/TestScripts/Proxies/httptrigger/run.csx
