+++ test/WebJobs.Script.Tests.Integration/TestScripts/Proxies/proxies.json
